Local DMA works
We have a DMA setup working that will use one shared buffer for RX and
TX. Tried triggering the RX SPI from the RXC IRQ, but that didn't work
out. So now we'll have four excess bytes in the output, which isn't
overly dramatic (we just need to ignore them later on).