Hardware update; does not work like this.
If the ignition remains on, I cannot get below 1 mA, possibly 2 mA (if I
put a 10 kOhm parallel resistor on the 12V to eliminate creep voltage).
This can never work. I need to figure out a way to reduce power
consumption during runtime and make the ignition switch a purely logical
device. This means a significant overhaul of the whole circuit.